circle id

+++ backend/migrations/20240815135549_circles.js
@@ -4,7 +4,7 @@
  */
 exports.up = function (knex) {
   return knex.schema.createTable("CIRCLES", function (table) {
-    table.bigInteger("CIRCLE_ID").primary(); // BIGINT primary key for circle ID
+    table.increments("CIRCLE_ID").primary(); // INT IDENTITY(1,1) primary key for circle ID
     table.string("CIRCLE_NAME", 100); // VARCHAR(100) for circle name
   });
 };
@@ -14,5 +14,5 @@ exports.up = function (knex) {
  * @returns { Promise<void> }
  */
 exports.down = function (knex) {
-  return knex.schema.dropTableIfExists("CIRCLES"); 
+  return knex.schema.dropTableIfExists("CIRCLES");
 };
